+++ src/pages/mission/droner/MissionReport.tsx
@@ -81,7 +81,7 @@ function MissionReport() {
       const tableList = []
       for (let i = 0; res.data.length > i; i++) {
         const table: any = {
-          updateAt: res.data[i]?.dronerTransactions?.[num -1]?.updateAt ?? res.data[i]?.updateAt,
+          updateAt: res.data[i]?.dronerTransactions?.[num - 1]?.updateAt ?? res.data[i]?.updateAt,
           name: res.data[i].firstname + ' ' + res.data[i].lastname,
           nickname: res.data[i].nickname,
           telephone: res.data[i].telephoneNo,