+++ src/resource/icon.ts
@@ -67,4 +67,7 @@ export default {
   comments_bullet: require('./media/images/comments-bullet.png'),
   play: require('./media/images/play.png'),
   pause: require('./media/images/pause.png'),
+  fertilizer: require('./media/images/fertilizer.png'),
+  account_cancel_disable: require('./media/images/account_cancel_disable.png'),
+  account_cancel: require('./media/images/account_cancel.png'),
 }